It turned out that Cramp::Action.render is protected so I needed to work around this problem by overriding the function in GameEventsAction.
The solution with a thin server running in a separate server turned out to be problematic WRT SSE so I stepped back to lifo's code that uses rack to exercise an app or a single action. It turns out that it's usable, I just need to handle sync response (e.g. 404) from HttpRouter in catch(:async) blocks after get. I suggested this to lifo and he said he'll add it to his test_helper.rb

I updated all specs to use the lifo's code (with a few changes) and made them all pass.