It's time to begin the work on story #4 ("Two players can simultaneously play two separate games"). I start by adding a failing scenario to game.feature.
To make it work, I had to play around with game_steps.rb a bit. I also changed the step that checks the contents of div#board so that it fails if the div contains more text besides of what we're looking for. The previous version succeeded when looking for "Lorem ipsum" even as div#board contained "Lorem ipsum dolor". Superfluous content was simply ignored.