The acceptance test for story #8 now passes plus all rspec tests while, obviously, the test for story #9 doesn't pass yet, but we're on the right track.
I also added "watch" links on the main page; while this is not used by cucumber helpers nor required by the acceptance test, it makes manual testing much easier. I haven't decided yet if the links will be visible in production or not.

So right now, we have the whole private/public game id business sorted out both in JS and in RB and I'm ready to start working on #9: the first turn is for the player who started the game.